feat(hum-mcp): carve the worker MCP HTTP bridge into a leaf crate
Move the axum JSON-RPC bridge (spawn_local_mcp / McpBridge / handle)
from nest-common into a standalone hum-mcp crate so remote worker
hives can expose tools to their compute over MCP without pulling the
daemon tree.

- New crate hum-mcp: mcp (protocol/catalogue/translate) + thrum-core
  (rid) + axum + tokio + parking_lot. Keeps the mcp crate pure (no
  server) — the bridge is the axum layer on top.
- hives/common/src/mcp_bridge.rs becomes a thin re-export shim for
  back-compat; existing nest_common::{spawn_local_mcp, McpBridge}
  call sites keep resolving.
- Moved the bridge tests (tools/list + tools/call round-trip) along.

Registered in workspace members; added to hives/common deps.
