+++ hives/common/Cargo.toml
@@ -9,6 +9,7 @@ description = "Shared building blocks for nests — regex Classifier for the dro
 hum-paths = { path = "../../hum-paths" }
 hum-identity = { path = "../../hum-identity" }
 hum-thrum = { path = "../../hum-thrum" }
+hum-mcp = { path = "../../hum-mcp" }
 drone = { path = "../../drone" }
 ensemble = { path = "../../ensemble" }
 mcp = { path = "../../mcp" }
